requirements.txt
Flask
flask_sqlalchemy
psycopg2-binary
flask-cors
flask-bcrypt